requirements.txt
# Core Data Science & ML
numpy==1.26.4
pandas==2.2.1
scikit-learn==1.4.1.post1
scipy==1.12.0

# Classical ML
xgboost==2.0.3

# Deep Learning
tensorflow==2.16.1
keras==3.0.5

# Visualization
matplotlib==3.8.3
seaborn==0.13.2

# Explainability
shap==0.45.0

# Imbalanced Data Handling
imbalanced-learn==0.12.0

# Configuration
pyyaml==6.0.1

# Jupyter / Experiments
jupyterlab==4.1.5
tqdm==4.66.2
ipykernel==6.29.3

# Dataset Download
kaggle==1.6.14

# Model Serialization
joblib==1.3.2
